List new listings if they are not already there
Now task will first check for the listings on the PoshUsers
account before deciding what to list.

+++ poshmark/tasks.py
@@ -97,10 +97,10 @@ def advanced_sharing(campaign_id):
     campaign = Campaign.objects.get(id=campaign_id)
     posh_user = campaign.posh_user
     logger = Log(logger_type=Log.CAMPAIGN, posh_user=posh_user)
-    listings = Listing.objects.filter(campaign__id=campaign_id)
+    campaign_listings = Listing.objects.filter(campaign__id=campaign_id)
     logged_hour_message = False
     fake_listing_titles = []
-    fake_listings_made = False
+    posted_new_listings = False
     max_deviation = round(campaign.delay / 2)
 
     campaign.status = '1'
@@ -113,29 +113,34 @@ def advanced_sharing(campaign_id):
         end_time = now + datetime.timedelta(days=1)
         while now < end_time and posh_user.status != PoshUser.INACTIVE and campaign.status == '1':
             while now.strftime('%I %p') in campaign.times and posh_user.status != PoshUser.INACTIVE and campaign.status == '1':
-                while len(fake_listing_titles) != len(listings):
-                    title = client.list_item()
-                    client.sleep(12)
-                    if client.check_listing(title):
-                        if client.share_item(title):
-                            fake_listing_titles.append(title)
-                        else:
-                            client.delete_listing(title)
-                    client.sleep(random.randint(16, 30))
+                if not posted_new_listings:
+                    listing_titles = client.get_all_listings()
+                    listings_to_list = [listing for listing in campaign_listings if listing.title not in listing_titles]
+                    while len(fake_listing_titles) != len(listings_to_list):
+                        title = client.list_item()
+                        client.sleep(12)
+                        if client.check_listing(title):
+                            if client.share_item(title):
+                                fake_listing_titles.append(title)
+                            else:
+                                client.delete_listing(title)
+                        client.sleep(random.randint(16, 30))
 
-                if not fake_listings_made:
-                    for index, value in enumerate(listings):
+                    for index, value in enumerate(listings_to_list):
                         client.update_listing(fake_listing_titles[index], value, 'Saks Fifth Avenue')
                         client.update_listing(value.title, value)
                         client.sleep(random.randint(16, 30))
-                    fake_listings_made = True
+                    posted_new_listings = True
 
                 campaign.refresh_from_db()
                 posh_user.refresh_from_db()
                 now = datetime.datetime.now(pytz.utc)
-                for listing in listings:
+
+                listing_titles = client.get_all_listings()
+
+                for listing_title in listing_titles:
                     pre_share_time = time.time()
-                    if client.share_item(listing.title):
+                    if client.share_item(listing_title):
                         positive_negative = 1 if random.random() < 0.5 else -1
                         deviation = random.randint(0, max_deviation) * positive_negative
                         post_share_time = time.time()
